% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{ACF}
\alias{ACF}
\alias{PACF}
\alias{CCF}
\title{(Partial) Autocorrelation and Cross-Correlation Function Estimation}
\usage{
ACF(.data, value = NULL, ...)

PACF(.data, value = NULL, ...)

CCF(.data, value1, value2, ...)
}
\arguments{
\item{.data}{A tsibble}

\item{value, value1, value2}{The column(s) from the tsibble to compute the ACF on}

\item{...}{Further arguments to be passed on to acf, pacf, and ccf}
}
\value{
The \code{Acf}, \code{Pacf} and \code{Ccf} functions return objects
of class "acf" as described in \code{\link[stats]{acf}} from the stats
package. The \code{taperedacf} and \code{taperedpacf} functions return
objects of class "mpacf".
}
\description{
The function \code{Acf} computes (and by default plots) an estimate of the
autocorrelation function of a (possibly multivariate) time series. Function
\code{Pacf} computes (and by default plots) an estimate of the partial
autocorrelation function of a (possibly multivariate) time series. Function
\code{Ccf} computes the cross-correlation or cross-covariance of two
univariate series.
}
\details{
The functions improve the \code{\link[stats]{acf}},
\code{\link[stats]{pacf}} and \code{\link[stats]{ccf}} functions. The main
differences are that \code{Acf} does not plot a spike at lag 0 when
\code{type=="correlation"} (which is redundant) and the horizontal axes show
lags in time units rather than seasonal units.

The tapered versions implement the ACF and PACF estimates and plots
described in Hyndman (2015), based on the banded and tapered estimates of
autocovariance proposed by McMurry and Politis (2010).
}
\examples{
elecdemand \%>\% ACF(Temperature)
elecdemand \%>\% PACF(Demand)
UKLungDeaths \%>\% CCF(mdeaths, fdeaths)
}
\references{
Hyndman, R.J. (2015). Discussion of ``High-dimensional
autocovariance matrices and optimal linear prediction''. \emph{Electronic
Journal of Statistics}, 9, 792-796.

McMurry, T. L., & Politis, D. N. (2010). Banded and tapered estimates for
autocovariance matrices and the linear process bootstrap. \emph{Journal of
Time Series Analysis}, 31(6), 471-482.
}
\seealso{
\code{\link[stats]{acf}}, \code{\link[stats]{pacf}},
\code{\link[stats]{ccf}}, \code{\link{tsdisplay}}
}
