% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ARIMA}
\alias{ARIMA}
\title{Estimate an ARIMA model}
\usage{
ARIMA(...)
}
\arguments{
\item{...}{Further arguments for arima}

\item{data}{A tsibble}

\item{formula}{Model specification.}

\item{stepwise}{Should stepwise be used?}

\item{greedy}{Should the stepwise search move to the next best option immediately?}

\item{approximation}{Should CSS be used during model selection?}
}
\description{
Estimate an ARIMA model
}
\examples{
# Manual ARIMA specification
USAccDeaths \%>\% as_tsibble \%>\% ARIMA(log(value) ~ pdq(0,1,1) + PDQ(0,1,1))

# Automatic ARIMA specification
tsibbledata::UKLungDeaths \%>\%
  ARIMA(log(mdeaths) ~ pdq(0,1,1) + PDQ(0,0,1) + fdeaths + fourier(K=4))

}
