% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima2.R
\name{ARIMA2}
\alias{ARIMA2}
\title{Estimate an ARIMA model}
\usage{
ARIMA2(data, formula, stepwise = TRUE, ...)
}
\arguments{
\item{data}{A tsibble}

\item{formula}{Model specification.}

\item{stepwise}{Should stepwise be used?}

\item{...}{Further arguments for arima}
}
\description{
Estimate an ARIMA model
}
\examples{
# Manual ARIMA specification
USAccDeaths \%>\% ARIMA2(log(value) ~ pdq(0,1,1) + PDQ(0,1,1))

# Automatic ARIMA specification
tsibbledata::UKLungDeaths \%>\% ARIMA2(log(mdeaths) ~ pdq(0,1,1) + PDQ(0,0,1) + fdeaths + fourier(K=4))

}
