% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{as_dable}
\alias{as_dable}
\title{Coerce a dataset to a dable}
\usage{
as_dable(data, decomposition)
}
\arguments{
\item{data}{A dataset containing a list decomposition column}

\item{decomposition}{A bare input containing the decomposition column's name}
}
\description{
Coerce a dataset to a dable
}
