% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{as_fable}
\alias{as_fable}
\title{Coerce a dataset to a fable}
\usage{
as_fable(data, model, forecast)
}
\arguments{
\item{data}{A dataset containing a list model column}

\item{model}{A bare input containing the model column's name}

\item{forecast}{A bare input containing the forecast column's name}
}
\description{
Coerce a dataset to a fable
}
