% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{as_mable}
\alias{as_mable}
\title{Coerce a dataset to a mable}
\usage{
as_mable(data, model)
}
\arguments{
\item{data}{A dataset containing a list model column}

\item{model}{A bare input containing the model column's name}
}
\description{
Coerce a dataset to a mable
}
