% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{biasadj}
\alias{biasadj}
\title{Bias adjust back-transformation functions}
\usage{
biasadj(bt_fn, fvar)
}
\arguments{
\item{bt_fn}{The back-transformation function}

\item{fvar}{The forecast variance}
}
\description{
To produce forecast means (instead of forecast medians) it is necessary to adjust the back-transformation function relative to the forecast variance.
}
\details{
More details about bias adjustment can be found in the transformations vignette: read the vignette:
\code{vignette("transformations", package = "fable")}
}
\examples{

adj_fn <- biasadj(exp, 1:10)
y <- rnorm(10)
exp(y)
adj_fn(y)

}
