% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{BoxCox}
\alias{BoxCox}
\alias{InvBoxCox}
\title{Box Cox Transformation}
\usage{
BoxCox(x, lambda)

InvBoxCox(x, lambda)
}
\arguments{
\item{x}{a numeric vector.}

\item{lambda}{a numeric value for the transformation parameter.}
}
\value{
a transformed numeric vector of the same length as x.
}
\description{
\code{BoxCox()} returns a transformation of the input variable using a Box-Cox
transformation. \code{InvBoxCox()} reverses the transformation.
}
\details{
The Box-Cox transformation is given by \deqn{f_\lambda(x) =\frac{x^\lambda -
1}{\lambda}}{f(x;lambda)=(x^lambda - 1)/lambda} if \eqn{\lambda\ne0}{lambda
is not equal to 0}. For \eqn{\lambda=0}{lambda=0},
\deqn{f_0(x)=\log(x)}{f(x;0)=log(x)}.
}
\examples{
library(tsibble)
airmiles \%>\% 
  as_tsibble() \%>\% 
  mutate(BoxCox = BoxCox(value, lambda = 0.3))

}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of
transformations. \emph{JRSS B} \bold{26} 211--246.
}
\author{
Rob J Hyndman & Mitchell O'Hara-Wild
}
