% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{breusch_godfrey}
\alias{breusch_godfrey}
\alias{breusch_godfrey.TSLM}
\title{Breusch-Godfrey Test}
\usage{
breusch_godfrey(x, ...)

\method{breusch_godfrey}{TSLM}(x, order = 1, type = c("Chisq", "F"), ...)
}
\arguments{
\item{x}{A model object to be tested.}

\item{...}{Further arguments for methods.}

\item{order}{The maximum order of serial correlation to test for.}

\item{type}{The type of test statistic to use.}
}
\description{
Breusch-Godfrey test for higher-order serial correlation.
}
\examples{

as_tsibble(USAccDeaths) \%>\%
  model(TSLM(value)) \%>\%
  hypothesize(breusch_godfrey)

}
\seealso{
\code{\link[lmtest:bgtest]{lmtest::bgtest()}}
}
