% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{common_periods}
\alias{common_periods}
\title{Extract frequencies for common seasonal periods}
\usage{
common_periods(x)
}
\arguments{
\item{x}{An object containing temporal data (such as a \code{tsibble}, \code{interval}, \code{datetime} and others.)}
}
\value{
A named vector of frequencies appropriate for the provided data.
}
\description{
Extract frequencies for common seasonal periods
}
\examples{
common_periods(tsibble::pedestrian)

}
\references{
\url{https://robjhyndman.com/hyndsight/seasonal-periods/}
}
