% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{components.ETS}
\alias{components.ETS}
\title{Extract estimated states from an ETS model.}
\usage{
\method{components}{ETS}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A \code{\link[fabletools:dable]{fabletools::dable()}} containing estimated states.
}
\description{
Extract estimated states from an ETS model.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(ets = ETS(log(value) ~ season("A"))) \%>\%
  components()
}
