% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{components}
\alias{components}
\title{Extract model or decomposition components}
\usage{
components(object, ...)
}
\arguments{
\item{object}{A model or decomposition}

\item{...}{Additional arguments passed to methods}
}
\description{
Extracts the decomposed components from an object, or the states from a state space model.
}
