% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CPPFABLEPostMean}
\alias{CPPFABLEPostMean}
\title{Compute FABLE posterior mean.}
\usage{
CPPFABLEPostMean(Y, gamma0, delta0sq, U_Y, V_Y, svalsY, kMax)
}
\arguments{
\item{Y}{Data matrix}

\item{gamma0}{Shape parameter for variance prior}

\item{delta0sq}{Rate parameter for variance prior}

\item{U_Y}{Left singular matrix}

\item{V_Y}{Right singular matrix}

\item{svalsY}{Singular values of Y}

\item{kMax}{Maximum of range of rank}
}
\value{
FABLE-posterior mean
}
\description{
Provides the explicitly available FABLE-posterior mean without any sampling
}
