% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CPPFABLESampler}
\alias{CPPFABLESampler}
\title{FABLE-posterior sampling with coverage-correction.}
\usage{
CPPFABLESampler(Y, gamma0, delta0sq, MC, U_Y, V_Y, svalsY, kEst, varInflation)
}
\arguments{
\item{Y}{Data matrix}

\item{gamma0}{Shape parameter for variance prior.}

\item{delta0sq}{Rate parameter for variance prior.}

\item{MC}{Desired number of Monte Carlo samples.}

\item{U_Y}{Left singular matrix}

\item{V_Y}{Right singular matrix}

\item{svalsY}{Singular values of Y}

\item{kEst}{Maximum of range of rank}

\item{varInflation}{Coverage-correction coefficient}
}
\description{
Provides independent FABLE-posterior samples of the covariance matrix
}
