% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CPPRankEstimator}
\alias{CPPRankEstimator}
\title{Estimates rank of factor model via JIC}
\usage{
CPPRankEstimator(Y, U_Y, V_Y, svalsY, kMax)
}
\arguments{
\item{Y}{n*p matrix of data; let r = min(n,p).}

\item{U_Y}{Left singular vectors (n*r).}

\item{V_Y}{Right singular vectors (p*r).}

\item{svalsY}{Singular values of Y (length r).}

\item{kMax}{Maximum rank to consider.}
}
\value{
Integer value of the estimated rank.
}
\description{
Implements the JIC rank estimation criterion.
}
