% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{dable}
\alias{dable}
\title{Create a new dable}
\usage{
dable(key_vals, data, decomposition)
}
\arguments{
\item{key_vals}{A list of values for keys}

\item{data}{The nested data (one row per decomposition)}

\item{decomposition}{A list of decompositions}
}
\description{
Create a new dable
}
