\name{FABLE-package}
\alias{FABLE-package}
\alias{FABLE}
\docType{package}
\title{
\packageTitle{FABLE}
}
\description{
\packageDescription{FABLE}
}
\details{
%The DESCRIPTION file:
\packageDESCRIPTION{FABLE}
\packageIndices{FABLE}
This package implements the Factor Analysis with BLEssing of dimension (FABLE) approach. The PseudoPosteriorSampler function provides pseudo-posterior samples from the coverage-corrected FABLE approach, while the PseudoPosteriorMean provides the pseudo-posterior mean.
}
\author{
\packageAuthor{FABLE}

Maintainer: \packageMaintainer{FABLE}
}
\references{
% ~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
% ~~ simple examples of the most important functions ~~
}
