% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABLE_wrapper.R
\name{FABLEPosteriorMean}
\alias{FABLEPosteriorMean}
\title{FABLE-posterior mean function}
\usage{
FABLEPosteriorMean(Y, gamma0 = 1, delta0sq = 1, maxProp = 0.5)
}
\arguments{
\item{Y}{Data matrix}

\item{gamma0}{Variance prior shape hyperparameter}

\item{delta0sq}{Variance prior rate hyperparameter}

\item{maxProp}{Upper bound on cumulative singular value proportion;
controls the range of the search space for the rank.}
}
\value{
List containing the estimate and other objects.
}
\description{
Description: Returns FABLE-posterior mean without sampling.
}
