% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABLE_wrapper.R
\name{FABLEPosteriorSampler}
\alias{FABLEPosteriorSampler}
\title{FABLE-posterior sampler function}
\usage{
FABLEPosteriorSampler(Y, gamma0 = 1, delta0sq = 1, maxProp = 0.95, MC = 1000)
}
\arguments{
\item{Y}{Data matrix}

\item{gamma0}{Variance prior shape hyperparameter}

\item{delta0sq}{Variance prior rate hyperparameter}

\item{maxProp}{Upper bound on cumulative singular value proportion;
controls the range of the search space for the rank.}

\item{MC}{Desired number of Monte Carlo samples}
}
\value{
List containing samples and other objects.
}
\description{
Description: Returns samples of the covariance from the FABLE-posterior.
}
