% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{fitted.AR}
\alias{fitted.AR}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{AR}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
as_tsibble(lh) \%>\%
  model(AR(value ~ order(3))) \%>\%
  fitted()
}
