% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croston.R
\name{fitted.croston}
\alias{fitted.croston}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{croston}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
poisson <- tsibble::tsibble(
  time = yearmonth("2012 Dec") + seq_len(24),
  count = rpois(24, lambda = 0.3)
)

poisson \%>\% 
  model(croston(count)) \%>\% 
  tidy()
}
