% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{fitted.TSGLM}
\alias{fitted.TSGLM}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{TSGLM}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(lm = TSGLM(log(value) ~ trend() + season())) \%>\%
  fitted()
}
