% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{fitted.VAR}
\alias{fitted.VAR}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{VAR}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Extracts the fitted values.
}
