% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecm.R
\name{fitted.VECM}
\alias{fitted.VECM}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{VECM}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
lung_deaths <- cbind(mdeaths, fdeaths) \%>\%
  as_tsibble(pivot_longer = FALSE)

lung_deaths \%>\%
  model(VAR(vars(mdeaths, fdeaths) ~ AR(3))) \%>\%
  fitted()
}
