% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecasting time series}
\usage{
forecast(object, h, ...)
}
\arguments{
\item{object}{a time series or time series model for which forecasts are
required}

\item{h}{Number of periods for forecasting}

\item{...}{Additional arguments affecting the forecasts produced.}
}
\value{
An object of class "\code{forecast}".
}
\description{
\code{forecast} is a generic function for forecasting from time series or
time series models. The function invokes particular \emph{methods} which
depend on the class of the first argument.
}
\author{
Rob J Hyndman
}
\keyword{ts}
