% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagwalk.R
\name{forecast.RW}
\alias{forecast.RW}
\title{Forecast a model from the fable package}
\usage{
\method{forecast}{RW}(
  object,
  new_data,
  specials = NULL,
  simulate = FALSE,
  bootstrap = FALSE,
  times = 5000,
  ...
)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{simulate}{If \code{TRUE}, prediction intervals are produced by simulation rather than using analytic formulae.}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{bootstrap = TRUE}.}

\item{...}{Other arguments passed to methods}
}
\value{
A list of forecasts.
}
\description{
Produces forecasts from a trained model.
}
\examples{
as_tsibble(Nile) \%>\%
  model(NAIVE(value)) \%>\%
  forecast()

library(tsibbledata)
aus_production \%>\%
  model(snaive = SNAIVE(Beer ~ lag("year"))) \%>\%
  forecast()
}
