% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbats.R
\name{forecast.TBATS}
\alias{forecast.TBATS}
\title{Forecast a model from the fable package}
\usage{
\method{forecast}{TBATS}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A list of forecasts.
}
\description{
Produces forecasts from a trained model.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(ets = ETS(log(value) ~ season("A"))) \%>\%
  forecast()
}
