% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arfima.R
\name{fracdiff}
\alias{fracdiff}
\title{Fractional differences}
\usage{
fracdiff(x, d)
}
\arguments{
\item{x}{Numeric vector to difference.}

\item{d}{Fractional differencing parameter.}
}
\value{
A numeric vector representing the fractional difference of \code{x}.
}
\description{
Computes the fractional difference of a numeric vector using the fast algorithm by Jensen and Nielsen (2014).
}
\references{
Jensen, Andreas Noack and Nielsen, Morten Ørregaard (2014) A Fast Fractional Difference Algorithm. Journal of Time Series Analysis 35(5), 428–436; \href{https://doi.org/10.1111/jtsa.12074}{doi:10.1111/jtsa.12074}.
}
\seealso{
\code{\link[fracdiff:diffseries]{fracdiff::diffseries()}} for the original function which additionally differences the data.
}
