% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arfima.R
\name{fracdiffinv}
\alias{fracdiffinv}
\title{Fractional Integration: Inverse of Fractional Differencing}
\usage{
fracdiffinv(x, d, xi = NULL)
}
\arguments{
\item{x}{Numeric vector to integrate.}

\item{d}{Fractional integration parameter.}

\item{xi}{A numeric vector containing the initial values for the fractional integral.}
}
\value{
A numeric vector representing the fractional integral of \code{x}.
}
\description{
Computes the inverse function of the fractional differencing function \code{\link[=fracdiff]{fracdiff()}}.
}
\examples{

s <- 1:10
d <- fracdiff(s, 0.5)
fracdiffinv(d, 0.5)
 
}
