% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R, R/lagwalk.R, R/lm.R, R/mean.R
\name{generate.ETS}
\alias{generate.ETS}
\alias{generate.RW}
\alias{generate.TSLM}
\alias{generate.model_mean}
\title{Generate new data from a fable model}
\usage{
\method{generate}{ETS}(x, new_data, bootstrap = FALSE, ...)

\method{generate}{RW}(x, new_data, bootstrap = FALSE, ...)

\method{generate}{TSLM}(x, new_data, specials, bootstrap = FALSE, ...)

\method{generate}{model_mean}(x, new_data, bootstrap = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{...}{Additional arguments for forecast model methods.}

\item{specials}{(passed by \code{\link[fablelite:forecast.mdl_df]{fablelite::forecast.mdl_df()}}).}
}
\description{
Simulates future paths from a dataset using a fitted model. Innovations are
sampled by the model's assumed error distribution. If \code{bootstrap} is \code{TRUE},
innovations will be sampled from the model's residuals. If \code{new_data}
contains the \code{.innov} column, those values will be treated as innovations.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(ETS(log(value) ~ season("A"))) \%>\% 
  generate(times = 100)
  
}
\seealso{
\code{\link[fablelite:generate.mdl_df]{fablelite::generate.mdl_df}}
}
