% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{generate.TSGLM}
\alias{generate.TSGLM}
\title{Generate new data from a fable model}
\usage{
\method{generate}{TSGLM}(x, new_data, specials, bootstrap = FALSE, ...)
}
\arguments{
\item{x}{A fitted model.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{...}{Other arguments passed to methods}
}
\description{
Simulates future paths from a dataset using a fitted model. Innovations are
sampled by the model's assumed error distribution. If \code{bootstrap} is \code{TRUE},
innovations will be sampled from the model's residuals. If \code{new_data}
contains the \code{.innov} column, those values will be treated as innovations.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(lm = TSGLM(log(value) ~ trend() + season())) \%>\%
  generate()
}
\seealso{
\code{\link[fabletools:generate.mdl_df]{fabletools::generate.mdl_df}}
}
