% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{glance.AR}
\alias{glance.AR}
\title{Glance a AR}
\usage{
\method{glance}{AR}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the AR model.
}
\details{
Contains the variance of residuals (\code{sigma2}), the log-likelihood (\code{log_lik}),
and information criterion (\code{AIC}, \code{AICc}, \code{BIC}).
}
\examples{
as_tsibble(lh) \%>\%
  model(AR(value ~ order(3))) \%>\%
  glance()
}
