% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{glance.fable_theta}
\alias{glance.fable_theta}
\title{Glance a theta method}
\usage{
\method{glance}{fable_theta}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the average method model.
}
\details{
Contains the variance of residuals (\code{sigma2}).
}
