% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{glance.model_mean}
\alias{glance.model_mean}
\title{Glance a average method model}
\usage{
\method{glance}{model_mean}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the average method model.
}
\details{
Contains the variance of residuals (\code{sigma2}).
}
\examples{
library(tsibbledata)
vic_elec \%>\%
  model(avg = MEAN(Demand)) \%>\%
  glance()
}
