% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbats.R
\name{glance.TBATS}
\alias{glance.TBATS}
\title{Glance a tbats model}
\usage{
\method{glance}{TBATS}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the tbats model.
}
