% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/lm.R
\name{interpolate.ARIMA}
\alias{interpolate.ARIMA}
\alias{interpolate.TSLM}
\title{Interpolate missing values from a fable model}
\usage{
\method{interpolate}{ARIMA}(object, new_data, specials, ...)

\method{interpolate}{TSLM}(object, new_data, specials, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{specials}{(passed by \code{\link[fablelite:forecast.mdl_df]{fablelite::forecast.mdl_df()}}).}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Applies a model specific estimation technique to predict the values of missing values in a \code{tsibble}, and replace them.
}
