% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{invert_transformation}
\alias{invert_transformation}
\title{Invert a transformation}
\usage{
invert_transformation(x, ...)
}
\arguments{
\item{x}{An expression/function which can be inverted}

\item{...}{Further arguments passed on to methods}
}
\value{
A object of class transformation
}
\description{
Can be used to invert a supported expression or transformation object for back-transforming.
}
\examples{
y <- rnorm(10)

# Works with expressions
my_trans <- invert_transformation(quote(log(y+1)))
my_trans

# Works with transformations
invert_transformation(my_trans)

}
