% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{knit_print.fable}
\alias{knit_print.fable}
\title{Method for printing in a \code{knitr} context.}
\usage{
\method{knit_print}{fable}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{fable}}.}

\item{...}{Further arguments passed on to \code{knitr::knit_print}.}
}
\description{
Method for printing in a \code{knitr} context.
}
