% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{LM}
\alias{LM}
\title{Fit a linear model with time series components}
\usage{
LM(data, formula, ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{Model specification.}

\item{...}{Additional arguments passed to lm}
}
\description{
Fit a linear model with time series components
}
\examples{

USAccDeaths \%>\% as_tsibble \%>\% LM(log(value) ~ trend() + season())
}
