% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{mable}
\alias{mable}
\title{Create a new mable}
\usage{
mable(data, model, parsed_model)
}
\arguments{
\item{data}{A tsibble used for modelling}

\item{model}{The fitted model}

\item{parsed_model}{The parsed model from \code{parse_model}}
}
\description{
Create a new mable
}
