% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{model_sum}
\alias{model_sum}
\title{Provide a succinct summary of a model}
\usage{
model_sum(x)
}
\arguments{
\item{x}{The model to summarise}
}
\description{
Similarly to pillar's type_sum and obj_sum, model_sum is used to provide brief model summaries.
}
