% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{multi_model}
\alias{multi_model}
\title{Multiple calls to a model for mass modelling}
\usage{
multi_model(data, cl, keys)
}
\arguments{
\item{data}{A tsibble}

\item{cl}{A modelling call}

\item{keys}{A set of keys to nest over}
}
\description{
Multiple calls to a model for mass modelling
}
