% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{multi_univariate}
\alias{multi_univariate}
\title{Multiple calls to a univariate model for each tsibble key}
\usage{
multi_univariate(data, cl)
}
\arguments{
\item{data}{A tsibble}

\item{cl}{A modelling call}
}
\description{
Multiple calls to a univariate model for each tsibble key
}
