% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{new_fable}
\alias{new_fable}
\title{Constructor}
\usage{
new_fable(x)
}
\arguments{
\item{x}{A fable-like object}
}
\description{
A constructor function for producing a fable (most useful for extension package authors)
}
