% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{new_fcdist}
\alias{new_fcdist}
\title{Create a forecast distribution object}
\usage{
new_fcdist(f, ..., transformation = ~.x, abbr = NULL)
}
\arguments{
\item{f}{A distribution function producing quantiles (such as \code{qnorm})}

\item{...}{Arguments for \code{f} function}

\item{transformation}{Transformation to be applied to resulting quantiles from \code{f}}

\item{abbr}{Abbreviation for display purposes, defaults to the object name of \code{f}}
}
\description{
Create a forecast distribution object
}
\examples{
mydist <- new_fcdist(qnorm, mean = rep(3, 10), sd = seq(0, 1, length.out=10),
 transformation = exp, abbr = "N")
mydist
hilo(mydist, 95)
}
