% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilo.R
\name{new_hilo}
\alias{new_hilo}
\title{Construct hilo intervals}
\usage{
new_hilo(lower, upper, level = NULL)
}
\arguments{
\item{lower, upper}{A numeric vector of values for lower and upper limits.}

\item{level}{Default \code{NULL} does not include 'level'. Otherwise values of
length 1 or as length of \code{lower}, expected between 0 and 100.}
}
\value{
A "hilo" object
}
\description{
Construct hilo intervals
}
\examples{
new_hilo(lower = rnorm(10), upper = rnorm(10) + 5, level = 95L)

}
\author{
Earo Wang & Mitchell O'Hara-Wild
}
