% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{new_mable}
\alias{new_mable}
\title{Constructor}
\usage{
new_mable(x)
}
\arguments{
\item{x}{A mable-like object}
}
\description{
A constructor function for producing a mable (most useful for extension package authors)
}
