% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{new_quantile}
\alias{new_quantile}
\title{Create a quantile object}
\usage{
new_quantile(f, ..., transformation = ~.x, abbr = NULL)
}
\arguments{
\item{f}{A quantile function}

\item{...}{Arguments for the quantile function}

\item{transformation}{Transformation to be applied to quantiles}

\item{abbr}{Abbreviation for display purposes, defaults to the name of the quantile function}
}
\description{
Create a quantile object
}
\examples{
qt <- new_quantile(qnorm, mean = rep(100, 100), sd = 1:100, transformation = log, abbr = "N")
qt
qt(0.5)
}
