% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specials.R
\name{new_specials_env}
\alias{new_specials_env}
\title{Create evaluation environment for specials}
\usage{
new_specials_env(..., .env = caller_env(), .required_specials = NULL,
  .bury = TRUE, .vals = NULL)
}
\arguments{
\item{...}{A named set of functions which used to parse formula inputs}

\item{.env}{The evaluation environment of the specials (to find other user objects)}

\item{.required_specials}{The names of specials which must be provided (and if not, are included used with no inputs).}

\item{.bury}{If TRUE, the specials are bound to a child environment of \code{.env}.}

\item{.vals}{A list of named values to be bound to the special functions}
}
\description{
Allows extension packages to make use of the formula parsing of specials.
}
