% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model}
\alias{parse_model}
\title{Parse the model specification for specials}
\usage{
parse_model(data, model, specials = NULL)
}
\arguments{
\item{data}{A dataset}

\item{model}{A validated model (from \code{validate_model})}

\item{specials}{A list of functions used to be evaluated from the formula.
If specials is NULL, no specials are computed}
}
\description{
Using a list of defined special functions, the user's formula specification and data
is parsed to extract important modelling components.
}
