% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model_lhs}
\alias{parse_model_lhs}
\title{Parse the RHS of the model formula for transformations}
\usage{
parse_model_lhs(model_lhs, data)
}
\arguments{
\item{model_lhs}{The expression for the lhs of the model (from \code{model_lhs()})}

\item{data}{Data to be used to find the response}
}
\description{
Parse the RHS of the model formula for transformations
}
