% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model_rhs}
\alias{parse_model_rhs}
\title{Parse the RHS of the model formula for specials}
\usage{
parse_model_rhs(model_rhs, specials)
}
\arguments{
\item{model_rhs}{The expression for the rhs of the model (from \code{model_rhs()})}

\item{specials}{The environment containing specials (from \code{new_specials_env()})}
}
\description{
Parse the RHS of the model formula for specials
}
