% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{possible_frequencies}
\alias{possible_frequencies}
\title{Extract potential time-series frequencies}
\usage{
possible_frequencies(x)
}
\arguments{
\item{x}{An object containing temporal data (such as a tsibble, interval, datetime and others.)}
}
\value{
A vector of possible frequencies appropriate for the time intervals provided.
}
\description{
Extract potential time-series frequencies
}
\examples{
possible_frequencies(tsibble::pedestrian)

}
\references{
\url{https://robjhyndman.com/hyndsight/seasonal-periods/}
}
