% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{print.fable}
\alias{print.fable}
\title{Print \code{fable} object.}
\usage{
\method{print}{fable}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{fable}}.}

\item{...}{Further arguments passed on to other \code{print} methods.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print \code{fable} object.
}
\details{
In an interactive context, the rendered table will be displayed in
a web browser. Otherwise, the HTML code will be printed as text.
}
