% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{refit.AR}
\alias{refit.AR}
\title{Refit an AR model}
\usage{
\method{refit}{AR}(object, new_data, specials = NULL, reestimate = FALSE, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{reestimate}{If \code{TRUE}, the coefficients for the fitted model will be re-estimated to suit the new data.}

\item{...}{Other arguments passed to methods}
}
\value{
A refitted model.
}
\description{
Applies a fitted AR model to a new dataset.
}
\examples{
lung_deaths_male <- as_tsibble(mdeaths)
lung_deaths_female <- as_tsibble(fdeaths)

fit <- lung_deaths_male \%>\%
  model(AR(value ~ 1 + order(10)))

report(fit)

fit \%>\%
  refit(lung_deaths_female) \%>\%
  report()
}
