% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{refit.ARIMA}
\alias{refit.ARIMA}
\title{Refit an ARIMA model}
\usage{
\method{refit}{ARIMA}(object, new_data, specials = NULL,
  reestimate = FALSE, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{specials}{(passed by \code{\link[fablelite:forecast.mdl_df]{fablelite::forecast.mdl_df()}}).}

\item{reestimate}{If \code{TRUE}, the coefficients for the fitted model will be re-estimated to suit the new data.}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Applies a fitted ARIMA model to a new dataset.
}
