% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{refit.ETS}
\alias{refit.ETS}
\title{Refit an ETS model}
\usage{
\method{refit}{ETS}(
  object,
  new_data,
  specials = NULL,
  reestimate = FALSE,
  reinitialise = TRUE,
  ...
)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{reestimate}{If \code{TRUE}, the coefficients for the fitted model will be re-estimated to suit the new data.}

\item{reinitialise}{If TRUE, the initial parameters will be re-estimated to suit the new data.}

\item{...}{Other arguments passed to methods}
}
\description{
Applies a fitted ETS model to a new dataset.
}
\examples{
lung_deaths_male <- as_tsibble(mdeaths)
lung_deaths_female <- as_tsibble(fdeaths)

fit <- lung_deaths_male \%>\%
  model(ETS(value))

report(fit)

fit \%>\%
  refit(lung_deaths_female, reinitialise = TRUE) \%>\%
  report()
}
