% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbats.R
\name{refit.TBATS}
\alias{refit.TBATS}
\title{Refit a tbats model}
\usage{
\method{refit}{TBATS}(object, new_data, specials = NULL, reestimate = FALSE, ...)
}
\description{
Applies a fitted tbats model to a new dataset.
}
