% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report information about an object}
\usage{
report(object, ...)
}
\arguments{
\item{object}{The object to report}

\item{...}{Additional options for the reporting function}
}
\description{
Displays the object in a suitable format for reporting.
}
