% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{residuals.AR}
\alias{residuals.AR}
\title{Extract residuals from a fable model}
\usage{
\method{residuals}{AR}(object, type = c("innovation", "regression"), ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{type}{The type of residuals to extract.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
as_tsibble(lh) \%>\%
  model(AR(value ~ order(3))) \%>\%
  residuals()
}
