% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croston.R
\name{residuals.croston}
\alias{residuals.croston}
\title{Tidy a fable model}
\usage{
\method{residuals}{croston}(object, ...)
}
\arguments{
\item{...}{Additional arguments to tidying method.}
}
\value{
The model's coefficients in a \code{tibble}.
}
\description{
Returns the coefficients from the model in a \code{tibble} format.
}
\examples{
poisson <- tsibble::tsibble(
  time = yearmonth("2012 Dec") + seq_len(24),
  count = rpois(24, lambda = 0.3)
)

poisson \%>\% 
  model(croston(count)) \%>\% 
  residuals()
}
