% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{residuals.ETS}
\alias{residuals.ETS}
\title{Extract residuals values from a fable model}
\usage{
\method{residuals}{ETS}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(ets = ETS(log(value) ~ season("A"))) \%>\%
  residuals()
}
