% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{residuals.fable_theta}
\alias{residuals.fable_theta}
\title{Extract residuals from a fable model}
\usage{
\method{residuals}{fable_theta}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
library(tsibbledata)
vic_elec \%>\%
  model(avg = MEAN(Demand)) \%>\%
  residuals()
}
