% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetar.R
\name{residuals.NNETAR}
\alias{residuals.NNETAR}
\title{Extract residuals values from a fable model}
\usage{
\method{residuals}{NNETAR}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
as_tsibble(airmiles) \%>\%
  model(nn = NNETAR(box_cox(value, 0.15))) \%>\%
  residuals()
}
