% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/ets.R, R/lm.R
\name{residuals.ARIMA}
\alias{residuals.ARIMA}
\alias{residuals.ETS}
\alias{residuals.TSLM}
\title{Extract residuals values from a fable model}
\usage{
\method{residuals}{ARIMA}(object, type = c("innovation", "regression"),
  ...)

\method{residuals}{ETS}(object, ...)

\method{residuals}{TSLM}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{type}{The type of the residuals to extract.}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Extracts the residuals.
}
