% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecm.R
\name{residuals.VECM}
\alias{residuals.VECM}
\title{Extract residuals from a fable model}
\usage{
\method{residuals}{VECM}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
lung_deaths <- cbind(mdeaths, fdeaths) \%>\%
  as_tsibble(pivot_longer = FALSE)

lung_deaths \%>\%
  model(VAR(vars(mdeaths, fdeaths) ~ AR(3))) \%>\%
  residuals()
}
