% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagwalk.R
\name{RW}
\alias{RW}
\alias{NAIVE}
\alias{SNAIVE}
\title{Random walk models}
\usage{
RW(data, formula = ~lag(1))

NAIVE(data, formula = ~lag(1))

SNAIVE(data, formula = ~lag("smallest"))
}
\arguments{
\item{data}{A data frame}

\item{formula}{Model specification.}
}
\description{
\code{RW()} returns a random walk model, which is equivalent to an ARIMA(0,1,0)
model with an optional drift coefficient included using \code{drift()}. \code{naive()} is simply a wrapper
to \code{rwf()} for simplicity. \code{snaive()} returns forecasts and
prediction intervals from an ARIMA(0,0,0)(0,1,0)m model where m is the
seasonal period.
}
\details{
The random walk with drift model is \deqn{Y_t=c + Y_{t-1} + Z_t}{Y[t]=c +
Y[t-1] + Z[t]} where \eqn{Z_t}{Z[t]} is a normal iid error. Forecasts are
given by \deqn{Y_n(h)=ch+Y_n}{Y[n+h]=ch+Y[n]}. If there is no drift (as in
\code{naive}), the drift parameter c=0. Forecast standard errors allow for
uncertainty in estimating the drift parameter (unlike the corresponding
forecasts obtained by fitting an ARIMA model directly).

The seasonal naive model is \deqn{Y_t= Y_{t-m} + Z_t}{Y[t]=Y[t-m] + Z[t]}
where \eqn{Z_t}{Z[t]} is a normal iid error.
}
\examples{
library(tsibbledata)
elecdemand \%>\% RW(Demand ~ drift())


Nile \%>\% as_tsibble \%>\% NAIVE

library(tsibbledata)
elecdemand \%>\% SNAIVE(Temperature ~ lag("day"))

}
