% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{scale_tsibble}
\alias{scale_tsibble}
\alias{scale_x_yearquarter}
\alias{scale_x_yearmonth}
\alias{scale_x_yearweek}
\title{tsibble datetime scales
This set of scales defines new scales for aggregate time structures defined in tsibble.}
\usage{
scale_x_yearquarter(...)

scale_x_yearmonth(...)

scale_x_yearweek(...)
}
\arguments{
\item{...}{Further arguments to be passed on to scale_x_date()}
}
\value{
A ggproto object inheriting from \code{Scale}
}
\description{
tsibble datetime scales
This set of scales defines new scales for aggregate time structures defined in tsibble.
}
