% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl.R
\name{STL}
\alias{STL}
\title{Multiple seasonal decomposition}
\usage{
STL(data, formula, ...)
}
\arguments{
\item{data}{A tsibble.}

\item{formula}{Decomposition specification.}

\item{...}{Other arguments passed to \code{\link[forecast]{mstl}}.}
}
\description{
Decompose a time series into seasonal, trend and remainder components.
Seasonal components are estimated iteratively using STL. Multiple seasonal periods are
allowed. The trend component is computed for the last iteration of STL.
Non-seasonal time series are decomposed into trend and remainder only.
In this case, \code{\link[stats]{supsmu}} is used to estimate the trend.
Optionally, the time series may be Box-Cox transformed before decomposition.
Unlike \code{\link[stats]{stl}}, \code{mstl} is completely automated.
}
\examples{
USAccDeaths \%>\% as_tsibble \%>\% STL(value ~ season(window = 10))
elecdemand \%>\% STL(Demand ~ season(period = "day"))
}
\seealso{
\code{\link[stats]{stl}}, \code{\link[stats]{supsmu}}
}
