% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream}
\alias{stream}
\title{Extend a fitted model with new data}
\usage{
stream(object, ...)
}
\arguments{
\item{object}{An object (such as a model) which can be extended with additional data}

\item{...}{Additional arguments passed on to other methods (usually streamed data and options)}
}
\description{
Extend a fitted model with new data
}
