% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbats.R
\name{TBATS}
\alias{TBATS}
\title{TBATS model (Exponential smoothing state space model with Box-Cox
transformation, ARMA errors, Trend and Seasonal components)}
\usage{
TBATS(formula, ...)
}
\description{
Fits a TBATS model as described in De Livera, Hyndman & Snyder (2011).
}
\references{
De Livera, A.M., Hyndman, R.J., & Snyder, R. D. (2011),
Forecasting time series with complex seasonal patterns using exponential
smoothing, \emph{Journal of the American Statistical Association},
\bold{106}(496), 1513-1527.
}
