% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{tidy.AR}
\alias{tidy.AR}
\title{Tidy a fable model}
\usage{
\method{tidy}{AR}(x)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
The model's coefficients in a \code{tibble}.
}
\description{
Returns the coefficients from the model in a \code{tibble} format.
}
\examples{
as_tsibble(lh) \%>\%
  model(AR(value ~ order(3))) \%>\%
  tidy()
}
