% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{tidy.ARIMA}
\alias{tidy.ARIMA}
\title{Tidy a fable model}
\usage{
\method{tidy}{ARIMA}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\description{
Returns the coefficients from the model in a tibble format.
}
