% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{tidy.ETS}
\alias{tidy.ETS}
\title{Tidy a fable model}
\usage{
\method{tidy}{ETS}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
The model's coefficients in a \code{tibble}.
}
\description{
Returns the coefficients from the model in a \code{tibble} format.
}
\examples{
as_tsibble(USAccDeaths) \%>\%
  model(ets = ETS(log(value) ~ season("A"))) \%>\%
  tidy()
}
