% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{tidy.model_mean}
\alias{tidy.model_mean}
\title{Tidy a fable model}
\usage{
\method{tidy}{model_mean}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
The model's coefficients in a \code{tibble}.
}
\description{
Returns the coefficients from the model in a \code{tibble} format.
}
\examples{
library(tsibbledata)
vic_elec \%>\%
  model(avg = MEAN(Demand)) \%>\%
  tidy()
}
