% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyforecast.R
\docType{package}
\name{tidyforecast-package}
\alias{tidyforecast}
\alias{tidyforecast-package}
\title{Forecasting Functions for Tidy Time Series}
\description{
Methods and tools for displaying and analysing univariate time series
forecasts in a tidy format including exponential smoothing via state space models and
automatic ARIMA modelling.
}
\details{
\tabular{ll}{ Package: \tab tidyforecast\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/tidyforecast}
  \item Report bugs at \url{https://github.com/tidyverts/tidyforecast/issues}
}

}
\author{
Rob J Hyndman

Maintainer: Rob.Hyndman@monash.edu
}
\keyword{package}
