% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.R
\name{traverse}
\alias{traverse}
\title{Recursively traverse an object}
\usage{
traverse(x, f = ~.x, g = ~.x, h = ~.x, base = ~is_syntactic_literal(.x)
  || is_symbol(.x))
}
\arguments{
\item{x}{The object to traverse}

\item{f}{A function (mapper) for combining the recursed components}

\item{g}{A function (mapper) applied to the object before recursion}

\item{h}{A function (mapper) applied to the base case}

\item{base}{The base case for the recursion}
}
\description{
Recursively traverse an object
}
