% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{TSLM}
\alias{TSLM}
\title{Fit a linear model with time series components}
\usage{
TSLM(data, formula, ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{Model specification.}

\item{...}{Additional arguments passed to lm}
}
\description{
Fit a linear model with time series components
}
\examples{

USAccDeaths \%>\% as_tsibble \%>\% TSLM(log(value) ~ trend() + season())

library(tsibbledata)
olympic_running \%>\% 
  TSLM(Time ~ trend()) \%>\% 
  interpolate(olympic_running)
}
