% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKLungDeaths}
\alias{UKLungDeaths}
\title{Monthly Deaths from Lung Diseases in the UK}
\format{Time series of class \code{tsibble}.}
\description{
Three time series giving the monthly deaths from bronchitis,
    emphysema and asthma in the UK, 1974--1979,
    both sexes (\code{ldeaths}), males (\code{mdeaths}) and
    females (\code{fdeaths}).
}
\examples{

UKLungDeaths

}
\keyword{datasets}
