% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{unitroot_options}
\alias{unitroot_options}
\title{Options for the unit root tests for order of integration}
\usage{
unitroot_options(ndiffs_alpha = 0.05, nsdiffs_alpha = 0.05,
  ndiffs_pval = ~feasts::unitroot_kpss(.)["kpss_pval"],
  nsdiffs_pval = ~feasts::stl_features(...)[2] < 0.64)
}
\arguments{
\item{ndiffs_alpha, nsdiffs_alpha}{The level for the test specified in the \code{pval} functions As long as \code{pval < alpha}, differences will be added.}

\item{ndiffs_pval, nsdiffs_pval}{A function (or lambda expression) which returns the probability of the . As long as \code{pval < alpha}, differences will be added.

Note: For the function for the seasonal p-value, the seasonal period will be provided as the \code{.period} argument to this function.}
}
\description{
Options for the unit root tests for order of integration
}
\seealso{
\link[feasts:unitroot_ndiffs]{feasts::unitroot_ndiffs} and \link[feasts:unitroot_nsdiffs]{feasts::unitroot_nsdiffs}
}
