% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{validate_model}
\alias{validate_model}
\title{Validate the user provided model}
\usage{
validate_model(model, data = NULL)
}
\arguments{
\item{model}{The user's model specification (unevaluated)}

\item{data}{A dataset used for automatic response selection}
}
\description{
Appropriately format the user's model for evaluation. Typically ran as one of the first steps
in a model function.
}
