% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{VAR}
\alias{VAR}
\title{Estimate an VAR model}
\usage{
VAR(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section).}

\item{...}{Further arguments for arima}
}
\description{
Estimate an VAR model
}
\section{Specials}{


\subsection{pdq}{
The \code{AR} special is used to specify the lag order for the auto-regression.
\preformatted{
AR(p = 0)
}

\tabular{ll}{
\code{p}        \tab The order of the auto-regressive (AR) terms.\cr
}
}

#' \subsection{xreg}{
Exogenous regressors can be included in an ARIMA model without explicitly using the \code{xreg()} special. Common exogenous regressor specials as specified in \code{\link{common_xregs}} can also be used. These regressors are handled using \code{\link[stats:model.frame]{stats::model.frame()}}, and so interactions and other functionality behaves similarly to \code{\link[stats:lm]{stats::lm()}}.
\preformatted{
xreg(...)
}

\tabular{ll}{
\code{...}      \tab Bare expressions for the exogenous regressors (such as \code{log(x)})
}
}
}

\examples{

\dontrun{
cbind(mdeaths, fdeaths) \%>\%
  as_tsibble(pivot_longer = FALSE) \%>\%
  model(VAR(vars(log(mdeaths), fdeaths) ~ AR(3)))
}

}
