% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ARIMA}
\alias{ARIMA}
\title{Fit ARIMA model to univariate time series}
\usage{
ARIMA(data, formula, period = "smallest", ic = c("aicc", "aic", "bic"),
  test = c("kpss", "adf", "pp"), seasonal.test = c("seas", "ocsb",
  "hegy", "ch"), ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{Model specification.}

\item{period}{Time series frequency for seasonal component.}

\item{ic}{The information criterion used to choose the best model
for auto.arima}

\item{test}{The unit root test used for selecting non-seasonal differences}

\item{seasonal.test}{The seasonal unit root test for selecting seasonal differences}

\item{...}{Further arguments for Arima or auto.arima}
}
\value{
See the \code{\link[stats]{arima}} function in the stats package.
The additional objects returned are \item{x}{The time series data}
\item{xreg}{The regressors used in fitting (when relevant).}
}
\description{
Largely a wrapper for the \code{\link[stats]{arima}} function in the stats
package. The main difference is that this function allows a drift term. It
is also possible to take an ARIMA model from a previous call to \code{Arima}
and re-apply it to the data \code{y}.
}
\details{
See the \code{\link[stats]{arima}} function in the stats package.
}
\examples{
# Manual ARIMA specification
USAccDeaths \%>\% ARIMA(log(value) ~ pdq(0,1,1) + PDQ(0,1,1))

# Automatic ARIMA specification
tsibbledata::UKLungDeaths \%>\% ARIMA(mdeaths ~ fdeaths)

}
\seealso{
\code{\link{auto.arima}}, \code{\link{forecast.Arima}}.
}
