% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ARIMA}
\alias{ARIMA}
\title{Estimate an ARIMA model}
\usage{
ARIMA(formula, stepwise = TRUE, greedy = TRUE, approximation = FALSE,
  ...)
}
\arguments{
\item{formula}{Model specification.}

\item{stepwise}{Should stepwise be used?}

\item{greedy}{Should the stepwise search move to the next best option immediately?}

\item{approximation}{Should CSS be used during model selection?}

\item{...}{Further arguments for arima}
}
\description{
Estimate an ARIMA model
}
\examples{
# Manual ARIMA specification
USAccDeaths \%>\% as_tsibble \%>\% 
  model(arima = ARIMA(log(value) ~ pdq(0,1,1) + PDQ(0,1,1)))

# Automatic ARIMA specification
tsibbledata::UKLungDeaths \%>\%
  model(arima = ARIMA(log(mdeaths) ~ pdq(0,1,1) + PDQ(0,0,1) + 
                      fdeaths + fourier(K=4)))

}
