% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{ETS}
\alias{ETS}
\title{Exponential smoothing state space model}
\usage{
ETS(data, formula, restrict = TRUE, ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{Model specification.}

\item{restrict}{If TRUE (default), the models with infinite variance will not
be allowed.}

\item{...}{Other arguments}
}
\value{
A \code{mable} containing fitted ETS models.

The generic accessor functions \code{fitted} and \code{residuals}
extract useful features of the value returned by \code{ETS} and associated
functions.
}
\description{
Returns ETS model specified by the formula.
}
\details{
Based on the classification of methods as described in Hyndman et al (2008).

The methodology is fully automatic. The model is chosen automatically if not
specified. This methodology performed extremely well on the M3-competition
data. (See Hyndman, et al, 2002, below.)
}
\examples{

USAccDeaths \%>\% as_tsibble \%>\% ETS(log(value) ~ season("A"))
}
\references{
Hyndman, R.J., Koehler, A.B., Snyder, R.D., and Grose, S. (2002)
"A state space framework for automatic forecasting using exponential
smoothing methods", \emph{International J. Forecasting}, \bold{18}(3),
439--454.

Hyndman, R.J., Akram, Md., and Archibald, B. (2008) "The admissible
parameter space for exponential smoothing models". \emph{Annals of
Statistical Mathematics}, \bold{60}(2), 407--426.

Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D. (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer-Verlag. \url{http://www.exponentialsmoothing.net}.
}
\seealso{
\code{\link[stats]{HoltWinters}}, \code{\link{RW}},
\code{\link{ARIMA}}.
}
\author{
Rob J Hyndman & Mitchell O'Hara-Wild
}
