% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{ETS}
\alias{ETS}
\title{Exponential smoothing state space model}
\usage{
ETS(formula, opt_crit = c("lik", "amse", "mse", "sigma", "mae"),
  nmse = 3, bounds = c("both", "usual", "admissible"), ic = c("aicc",
  "aic", "bic"), restrict = TRUE, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{opt_crit}{The optimization criterion. Defaults to the log-likelihood
\code{"lik"}, but can also be set to \code{"mse"} (Mean Square Error), \code{"amse"}
(Average MSE over first \code{nmse} forecast horizons), \code{"sigma"} (Standard
deviation of residuals), or \code{"mae"} (Mean Absolute Error).}

\item{nmse}{If \code{opt_crit == "amse"}, \code{nmse} provides the number of steps for
average multistep MSE (\code{1<=nmse<=30}).}

\item{bounds}{Type of parameter space to impose: \code{"usual"} indicates
all parameters must lie between specified lower and upper bounds;
\code{"admissible"} indicates parameters must lie in the admissible space;
\code{"both"} (default) takes the intersection of these regions.}

\item{ic}{The information criterion used in selecting the model.}

\item{restrict}{If TRUE (default), the models with infinite variance will not
be allowed.}

\item{...}{Other arguments}
}
\description{
Returns ETS model specified by the formula.
}
\details{
Based on the classification of methods as described in Hyndman et al (2008).

The methodology is fully automatic. The model is chosen automatically if not
specified. This methodology performed extremely well on the M3-competition
data. (See Hyndman, et al, 2002, below.)
}
\examples{

USAccDeaths \%>\% as_tsibble \%>\% model(ETS(log(value) ~ season("A")))
}
\references{
Hyndman, R.J., Koehler, A.B., Snyder, R.D., and Grose, S. (2002)
"A state space framework for automatic forecasting using exponential
smoothing methods", \emph{International J. Forecasting}, \bold{18}(3),
439--454.

Hyndman, R.J., Akram, Md., and Archibald, B. (2008) "The admissible
parameter space for exponential smoothing models". \emph{Annals of
Statistical Mathematics}, \bold{60}(2), 407--426.

Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D. (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer-Verlag. \url{http://www.exponentialsmoothing.net}.
}
\seealso{
\code{\link[stats]{HoltWinters}}, \code{\link{RW}},
\code{\link{ARIMA}}.
}
\author{
Rob J Hyndman & Mitchell O'Hara-Wild
}
