\name{FABLE-package}
\alias{FABLE-package}
\alias{FABLE}
\docType{package}
\title{
\packageTitle{FABLE}
}
\description{
\packageDescription{FABLE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FABLE}
\packageIndices{FABLE}
}
\author{
\packageAuthor{FABLE}

Maintainer: \packageMaintainer{FABLE}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
