\name{FABLE-package}
\alias{FABLE-package}
\alias{FABLE}
\docType{package}
\title{
\packageTitle{FABLE}
}
\description{
\packageDescription{FABLE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FABLE}
\packageIndices{FABLE}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FABLE}

Maintainer: \packageMaintainer{FABLE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
