% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\docType{package}
\name{fable-package}
\alias{fable}
\alias{fable-package}
\title{Forecasting Functions for Tidy Time Series}
\description{
Methods and tools for displaying and analysing univariate time series
forecasts in a tidy format including exponential smoothing via state space models and
automatic ARIMA modelling.
}
\details{
\tabular{ll}{ Package: \tab fable\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/fable}
  \item Report bugs at \url{https://github.com/tidyverts/fable/issues}
}

}
\author{
Rob J Hyndman, Mitchell O'Hara-Wild, Earo Wang

Maintainer: Rob.Hyndman@monash.edu
}
\keyword{package}
