% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable-package.R
\docType{package}
\name{fable-package}
\alias{fable}
\alias{fable-package}
\title{fable: Forecasting tidy time series}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The R package *fable* provides methods and tools
    for displaying and analysing time series forecasts. Data, model and
    forecast objects are all stored in a tidy format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fable.tidyverts.org}
  \item Report bugs at \url{https://github.com/tidyverts/fable/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

}
\keyword{internal}
