% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable-package.R
\docType{package}
\name{fable-package}
\alias{fable}
\alias{fable-package}
\title{fable: Forecasting Models for Tidy Time Series}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a collection of commonly used univariate and multivariate
    time series forecasting models including automatically selected exponential 
    smoothing (ETS) and autoregressive integrated moving average (ARIMA) models.
    These models work within the 'fable' framework provided by the 'fabletools'
    package, which provides the tools to evaluate, visualise, and combine models 
    in a workflow consistent with the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fable.tidyverts.org}
  \item \url{https://github.com/tidyverts/fable}
  \item Report bugs at \url{https://github.com/tidyverts/fable/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

Other contributors:
\itemize{
  \item Gabriel Caceres (NNETAR implementation) [contributor]
}

}
\keyword{internal}
