% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABLE-package.R
\docType{package}
\name{FABLE}
\alias{FABLE-package}
\alias{FABLE}
\title{FABLE: Factor Analysis with BLEssing of dimensionality}
\description{
FABLE provides exact Monte Carlo samples of the covariance matrix in Bayesian factor analysis.
This package has functions for obtaining posterior mean and posterior samples of the covariance matrix.
The posterior mean function is \code{\link{CPPFABLEPostMean}}.
The posterior sampler function is \code{\link{CPPFABLESampler}}.
}
\references{
Chattopadhyay, S., Zhang, A. R., & Dunson, D. B. (2024). Blessing of dimension in Bayesian inference on covariance matrices. arXiv preprint arXiv:2404.03805.
}
\author{
Shounak Chattopadhyay
}
\keyword{Embarrassingly}
\keyword{approximation.}
\keyword{parallel,}
\keyword{posterior}
