% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a new fable}
\usage{
fable(key_vals, data, model, forecast)
}
\arguments{
\item{key_vals}{A set of key values}

\item{data}{A tsibble used for modelling}

\item{model}{The fitted model}

\item{forecast}{A list of tsibble forecasts}
}
\description{
Create a new fable
}
