% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a new fable}
\usage{
fable(key_vals, data, model, forecast)
}
\arguments{
\item{data}{The nested data (one row per model)}

\item{model}{A list of models}

\item{forecast}{A list of tsibble forecasts}
}
\description{
Create a new fable
}
