% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{fitted.model_mean}
\alias{fitted.model_mean}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{model_mean}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
library(tsibbledata)
vic_elec \%>\%
  model(avg = MEAN(Demand)) \%>\%
  fitted()
}
