% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetar.R
\name{fitted.NNETAR}
\alias{fitted.NNETAR}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{NNETAR}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
as_tsibble(airmiles) \%>\%
  model(nn = NNETAR(box_cox(value, 0.15))) \%>\%
  fitted()
}
