% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/ets.R, R/lagwalk.R, R/lm.R, R/mean.R
\name{fitted.ARIMA}
\alias{fitted.ARIMA}
\alias{fitted.ETS}
\alias{fitted.RW}
\alias{fitted.TSLM}
\alias{fitted.model_mean}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{ARIMA}(object, ...)

\method{fitted}{ETS}(object, ...)

\method{fitted}{RW}(object, ...)

\method{fitted}{TSLM}(object, ...)

\method{fitted}{model_mean}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Extracts the fitted values.
}
