% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/ets.R
\name{fitted.ARIMA}
\alias{fitted.ARIMA}
\alias{fitted.ETS}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{ARIMA}(object, ...)

\method{fitted}{ETS}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Extracts the fitted values.
}
