% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetar.R
\name{forecast.NNETAR}
\alias{forecast.NNETAR}
\title{Forecast a model from the fable package}
\usage{
\method{forecast}{NNETAR}(object, new_data, specials = NULL,
  bootstrap = FALSE, times = 1000, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{specials}{(passed by \code{\link[fablelite:forecast.mdl_df]{fablelite::forecast.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{boostrap = TRUE}.}

\item{...}{Additional arguments for forecast model methods.}
}
\description{
Produces forecasts from a trained model.
}
