% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R, R/ets.R, R/lm.R
\name{forecast.ARIMA}
\alias{forecast.ARIMA}
\alias{forecast.ETS}
\alias{forecast.TSLM}
\title{Forecast a model from the fable package}
\usage{
\method{forecast}{ARIMA}(object, new_data = NULL, specials = NULL,
  bootstrap = FALSE, times = 5000, ...)

\method{forecast}{ETS}(object, new_data, specials = NULL,
  simulate = FALSE, bootstrap = FALSE, times = 5000, ...)

\method{forecast}{TSLM}(object, new_data, specials = NULL,
  bootstrap = FALSE, times = 5000, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{specials}{(passed by \code{\link[fablelite:forecast.mdl_df]{fablelite::forecast.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{boostrap = TRUE}.}

\item{...}{Additional arguments for forecast model methods.}

\item{simulate}{If \code{TRUE}, prediction intervals are produced by simulation rather than using analytic formulae.}
}
\description{
Produces forecasts from a trained model.
}
