% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilo.R
\name{lower}
\alias{lower}
\alias{upper}
\alias{level}
\alias{is_hilo}
\title{Helpers for "hilo"}
\usage{
lower(x)

upper(x)

level(x)

is_hilo(x)
}
\arguments{
\item{x}{A "hilo" object.}
}
\description{
Helpers for "hilo"
}
