% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{mable}
\alias{mable}
\title{Create a new mable}
\usage{
mable(key_vals, data, model)
}
\arguments{
\item{data}{The nested data (one row per model)}

\item{model}{A list of models}
}
\description{
Create a new mable
}
