% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specials.R
\name{new_specials_env}
\alias{new_specials_env}
\title{Create evaluation environment for specials}
\usage{
new_specials_env(..., parent_env = caller_env(), required_specials = NULL)
}
\arguments{
\item{...}{A named set of functions which used to parse formula inputs}

\item{parent_env}{The parent environment of the specials (to find other user objects)}

\item{required_specials}{The names of specials which must be provided (and if not, are included used with no inputs).}
}
\description{
Allows extension packages to make use of the formula parsing of specials.
}
