% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model_rhs}
\alias{parse_model_rhs}
\title{Parse the RHS of the model formula for specials}
\usage{
parse_model_rhs(model_rhs, data, specials = NULL)
}
\arguments{
\item{model_rhs}{The expression for the rhs of the model (from \code{model_rhs()})}

\item{data}{Data provided by the user to evaluate the special functions within}

\item{specials}{The environment containing specials (from \code{new_specials_env()}).
If specials is NULL, no specials are computed}
}
\description{
Parse the RHS of the model formula for specials
}
