% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tsibble}
\alias{forecast}
\alias{model}
\alias{imitate}
\alias{accuracy}
\alias{box_cox}
\alias{interpolate}
\alias{stream}
\alias{refit}
\alias{report}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fablelite}{\code{\link[fablelite]{forecast}}, \code{\link[fablelite]{model}}, \code{\link[fablelite]{imitate}}, \code{\link[fablelite]{accuracy}}, \code{\link[fablelite]{box_cox}}, \code{\link[fablelite]{interpolate}}, \code{\link[fablelite]{stream}}, \code{\link[fablelite]{refit}}, \code{\link[fablelite]{report}}}

  \item{tsibble}{\code{\link[tsibble]{\%>\%}}, \code{\link[tsibble]{as_tsibble}}}
}}

