% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{as_tsibble}
\alias{reexports}
\alias{type_sum}
\alias{reexports}
\alias{obj_sum}
\alias{reexports}
\alias{autoplot}
\alias{reexports}
\alias{autolayer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{autolayer}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{pillar}{\code{\link[pillar]{type_sum}}, \code{\link[pillar]{obj_sum}}}

  \item{tsibble}{\code{\link[tsibble]{as_tsibble}}}
}}

