% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{residuals.model_mean}
\alias{residuals.model_mean}
\title{Extract residuals values from a fable model}
\usage{
\method{residuals}{model_mean}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
library(tsibbledata)
vic_elec \%>\%
  model(avg = MEAN(Demand)) \%>\%
  residuals()
}
