% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{TSLM}
\alias{TSLM}
\title{Fit a linear model with time series components}
\usage{
TSLM(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Additional arguments passed to lm}
}
\description{
Fit a linear model with time series components
}
\section{Specials}{


\subsection{xreg}{
Exogenous regressors can be included in an ARIMA model without explicitly using the \code{xreg()} special. Common exogenous regressor specials as specified in \code{\link{common_xregs}} can also be used. These regressors are handled using \code{\link[stats:model.frame]{stats::model.frame()}}, and so interactions and other functionality behaves similarly to \code{\link[stats:lm]{stats::lm()}}.
\preformatted{
xreg(...)
}

\tabular{ll}{
\code{...}      \tab Bare expressions for the exogenous regressors (such as \code{log(x)})
}
}
}

\examples{

USAccDeaths \%>\% as_tsibble \%>\% model(TSLM(log(value) ~ trend() + season()))

library(tsibbledata)
olympic_running \%>\% 
  model(TSLM(Time ~ trend())) \%>\% 
  interpolate(olympic_running)
}
